Handle keyboard events for WebContents that do not have owner window
